+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.33] — 2026-07-07
+
 ### Added
 
   - Interactive type queries (`typeof`) and symbol resolution (`resolve`) are now